Moving create-spell file outside of the index.js file.
